initial commit
>A kitsune is a shapeshifter, and usually when it reaches the age of 100
years, it learn the ability to take on a human form. Thus, they have to
be a fox for a hundred years before it can shapeshift from a fox to a
human and back again. It is also said that a kitsune can duplicate other
human beings, in other words shapeshift into the look-a-likes of
different people.

-- http://www.mythicalcreaturesguide.com/page/Kitsune